+++ tests/CodeigniterModelTest.php
@@ -85,7 +85,7 @@ class CodeigniterModelTest extends \PHPUnit_Framework_TestCase
 
         $comments = $this->ci->post->with('user')->all();
 
-        $this->assertInstanceOf('User', $comments[0]->user);
+        $this->assertEquals(1, $comments[0]->user->id);
     }
 
     /**