+++ CHANGELOG.md
@@ -9,6 +9,9 @@ All notable changes to `Wildfire` will be documented in this file.
 - `$belongs_to` for one-to-one relationships
 - `$hidden` for hiding columns
 
+### Changed
+- Version of `rougin/describe` to `~1.5`
+
 ## [0.3.1](https://github.com/rougin/wildfire/compare/v0.3.0...v0.3.1) - 2016-09-12
 
 ### Added